# tests/EventBooking.Web.Tests/Contracts/openapi-v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "EventBooking API",
    "description": "EventBooking staff API and anonymous Attendee booking links.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:5001/"
    }
  ],
  "paths": {
...
    "/api/events/{id}": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "Read one event.",
        "description": "Reads one event with its capacities, filtered the same way the list is. An event outside the caller's scope reads as not found.",
        "operationId": "getEvent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          { }
        ],
        "x-mcp-tool": "get_event",
        "x-agent-hints": {
          "readOnly": true,
          "destructive": false,
          "idempotent": true,
          "openWorld": false
        }
      }
    },
...
    "/api/attendees/{id}": {
      "put": {
        "tags": [
          "Attendees"
        ],
        "summary": "Update an attendee.",
        "description": "Updates an attendee's name, email and group (FR-4.2). A group change that would alter an active booking's requirements is refused.",
        "operationId": "updateAttendee",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveAttendeeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          { }
        ],
        "x-capability": "ManageAttendees",
        "x-mcp-tool": "update_attendee",
        "x-agent-hints": {
          "readOnly": false,
          "destructive": true,
          "idempotent": true,
          "openWorld": false
        }
      },
      "delete": {
        "tags": [
          "Attendees"
        ],
        "summary": "Delete an attendee.",
        "description": "Deletes an attendee and their bookings (FR-4.5). Two-step: without confirm=true the call reports its consequence and changes nothing.",
        "operationId": "deleteAttendee",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "confirm",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          { }
        ],
        "x-capability": "ManageAttendees",
        "x-mcp-tool": "delete_attendee",
        "x-agent-hints": {
          "readOnly": false,
          "destructive": true,
          "idempotent": true,
          "openWorld": false
        }
      }
    },
...
    "/api/audit": {
      "get": {
        "tags": [
          "Audit"
        ],
        "summary": "Search the audit log.",
        "description": "Searches audit entries, scoped to the buckets the caller's capabilities reach (FR-12.3), one keyset page at a time.",
        "operationId": "searchAudit",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "action",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "actorType",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "actorId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entityType",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entityId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageOfAuditRowResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          { }
        ],
        "x-eventbooking-list": true,
        "x-mcp-tool": "search_audit",
        "x-agent-hints": {
          "readOnly": true,
          "destructive": false,
          "idempotent": true,
          "openWorld": false
        }
      }
    },
    "/api/audit/attendees/{id}": {
      "get": {
        "tags": [
          "Audit"
        ],
        "summary": "Read an attendee's history.",
        "description": "Reads the audit history of one attendee.",
        "operationId": "getAttendeeAuditHistory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageOfAuditRowResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          { }
        ],
        "x-eventbooking-list": true,
        "x-capability": "ViewAttendeeAudit",
        "x-mcp-tool": "get_attendee_audit_history",
        "x-agent-hints": {
          "readOnly": true,
          "destructive": false,
          "idempotent": true,
          "openWorld": false
        }
      }
    },
    "/api/audit/events/{id}": {
      "get": {
        "tags": [
          "Audit"
        ],
        "summary": "Read an event's history.",
        "description": "Reads the audit history of one event, including its proposal.",
        "operationId": "getEventAuditHistory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageOfAuditRowResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          { }
        ],
        "x-eventbooking-list": true,
        "x-capability": "ViewEventAudit",
        "x-mcp-tool": "get_event_audit_history",
        "x-agent-hints": {
          "readOnly": true,
          "destructive": false,
          "idempotent": true,
          "openWorld": false
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiLink": {
        "required": [
          "href",
          "method",
          "operationId"
        ],
        "type": "object",
        "properties": {
          "href": {
            "type": "string"
          },
          "method": {
            "type": "string"
          },
          "operationId": {
            "type": "string"
          }
        }
      },
...
      "AuditRowResponse": {
        "required": [
          "id",
          "timestamp",
          "entityType",
          "entityId",
          "action",
          "actorType",
          "actorId",
          "actorDisplay",
          "details",
          "_links"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "entityType": {
            "type": "string"
          },
          "entityId": {
            "type": "string",
            "format": "uuid"
          },
          "action": {
            "type": "string"
          },
          "actorType": {
            "type": "string"
          },
          "actorId": {
            "type": [
              "null",
              "string"
            ]
          },
          "actorDisplay": {
            "type": [
              "null",
              "string"
            ]
          },
          "details": {
            "type": [
              "null",
              "string"
            ]
          },
          "_links": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/ApiLink"
            }
          }
        }
      },
...
      "EventCapacityResponse": {
        "required": [
          "appointmentTypeId",
          "code",
          "name",
          "totalHeadcount",
          "remainingCapacity",
          "_links"
        ],
        "type": "object",
        "properties": {
          "appointmentTypeId": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "totalHeadcount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "remainingCapacity": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "_links": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/ApiLink"
            }
          }
        }
      },
...
      "EventResponse": {
        "required": [
          "id",
          "proposalId",
          "locationId",
          "locationCode",
          "locationName",
          "time",
          "status",
          "capacities",
          "activeBookings",
          "_links"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "proposalId": {
            "type": "string",
            "format": "uuid"
          },
          "locationId": {
            "type": "string",
            "format": "uuid"
          },
          "locationCode": {
            "type": "string"
          },
          "locationName": {
            "type": "string"
          },
          "time": {
            "$ref": "#/components/schemas/EventTimeResponse"
          },
          "status": {
            "type": "string"
          },
          "capacities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventCapacityResponse"
            }
          },
          "activeBookings": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "_links": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/ApiLink"
            }
          }
        }
      },
      "EventTimeResponse": {
        "required": [
          "date",
          "startTime",
          "durationMinutes",
          "startLocal",
          "endLocal",
          "startUtc",
          "endUtc",
          "timeZoneId",
          "zoneAbbreviation"
        ],
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "startTime": {
            "type": "string",
            "format": "time"
          },
          "durationMinutes": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "startLocal": {
            "type": "string",
            "format": "date-time"
          },
          "endLocal": {
            "type": "string",
            "format": "date-time"
          },
          "startUtc": {
            "type": "string",
            "format": "date-time"
          },
          "endUtc": {
            "type": "string",
            "format": "date-time"
          },
          "timeZoneId": {
            "type": "string"
          },
          "zoneAbbreviation": {
            "type": "string"
          }
        }
      },
...
      "PageOfAuditRowResponse": {
        "required": [
          "items",
          "nextCursor"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditRowResponse"
            }
          },
          "nextCursor": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "SaveAttendeeRequest": {
        "required": [
          "name",
          "email",
          "attendeeGroupId"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "email": {
            "type": [
              "null",
              "string"
            ]
          },
          "attendeeGroupId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}